Also remove ability for squashed transplant
