+++ servers/services/src/testFixtures/java/org/projectnessie/services/impl/AbstractTestAccessChecks.java
@@ -236,7 +236,6 @@ public abstract class AbstractTestAccessChecks extends BaseTestServiceImpl {
             null,
             asList(source1.getHash(), source.getHash()),
             source1.getName(),
-            true,
             emptyList(),
             NORMAL,
             false,
@@ -390,7 +389,6 @@ public abstract class AbstractTestAccessChecks extends BaseTestServiceImpl {
                           null,
                           singletonList(ref.getHash()),
                           ref.getName(),
-                          false,
                           emptyList(),
                           NORMAL,
                           false,
