+++ versioned/tests/src/main/java/org/projectnessie/versioned/tests/AbstractNamespaceValidation.java
@@ -204,15 +204,12 @@ public abstract class AbstractNamespaceValidation extends AbstractNestedVersionS
   }
 
   enum NamespaceValidationMergeTransplant {
-    MERGE(true, false, false, false, false),
-    MERGE_CREATE(true, true, false, false, false),
-    MERGE_DELETE(true, false, true, false, true),
-    TRANSPLANT_SQUASH(false, false, false, false, false),
-    TRANSPLANT_INDIVIDUAL(false, false, false, true, false),
-    TRANSPLANT_CREATE_SQUASH(true, true, false, false, false),
-    TRANSPLANT_CREATE_INDIVIDUAL(true, true, false, true, false),
-    TRANSPLANT_DELETE_SQUASH(false, false, true, false, true),
-    TRANSPLANT_DELETE_INDIVIDUAL(false, false, true, true, true),
+    MERGE(true, false, false, false),
+    MERGE_CREATE(true, true, false, false),
+    MERGE_DELETE(true, false, true, true),
+    TRANSPLANT(false, false, false, false),
+    TRANSPLANT_CREATE(true, true, false, false),
+    TRANSPLANT_DELETE(false, false, true, true),
     ;
 
     /** Whether to merge (or transplant, if false). */
@@ -227,21 +224,16 @@ public abstract class AbstractNamespaceValidation extends AbstractNestedVersionS
      */
     final boolean deleteNamespaceOnTarget;
 
-    /** Whether merge/transplant shall keep individual commits or "squash" those. */
-    final boolean individualCommits;
-
     final boolean error;
 
     NamespaceValidationMergeTransplant(
         boolean merge,
         boolean createNamespaceOnTarget,
         boolean deleteNamespaceOnTarget,
-        boolean individualCommits,
         boolean error) {
       this.merge = merge;
       this.createNamespaceOnTarget = createNamespaceOnTarget;
       this.deleteNamespaceOnTarget = deleteNamespaceOnTarget;
-      this.individualCommits = individualCommits;
       this.error = error;
     }
   }
@@ -326,7 +318,6 @@ public abstract class AbstractNamespaceValidation extends AbstractNestedVersionS
                             .toBranch(root)
                             .addSequenceToTransplant(
                                 commit1.getCommitHash(), commit2.getCommitHash())
-                            .keepIndividualCommits(mode.individualCommits)
                             .build());
 
     if (mode.deleteNamespaceOnTarget) {
